refactor: lazy and eager both defined in base.py rather than register.py
