+++ src/eval_framework/tasks/base.py
@@ -1,3 +1,4 @@
+import importlib
 import logging
 import os
 import random
@@ -13,7 +14,7 @@ from datasets import DatasetDict, DownloadConfig, load_dataset
 from eval_framework.shared.types import BaseMetricContext, Completion, Error, RawCompletion
 from eval_framework.tasks.dataset_revisions import pinned_revision
 from eval_framework.tasks.markdown_doc import markdown_doc as render_markdown_doc
-from eval_framework.tasks.task import ResponseType, Sample, Task
+from eval_framework.tasks.task import EvalFactory, ResponseType, Sample, Task
 from eval_framework.tasks.utils import classproperty, raise_errors
 from template_formatting.formatter import BaseFormatter, Message, Role
 
@@ -501,3 +502,125 @@ def resolve_overwrite_subjects[SubjectType](
             )
 
     return chosen_subjects
+
+
+class Lazy(EvalFactory):
+    """
+    Create eval from qualified class path; Delays importing modules until
+    eval is constructed.
+    """
+
+    def __init__(self, class_name: str, module: str) -> None:
+        """
+        Args:
+            class_name: The name of the task class to import.
+            module: The module to import the task class from.
+        """
+        self._class_name = class_name
+        self._module = module
+        self._loaded: type[BaseTask] | None = None
+
+    @property
+    def source_module(self) -> str:
+        return self._module
+
+    def id(self) -> str:
+        return self._class_name
+
+    def task_class(self) -> type[BaseTask]:
+        if self._loaded is None:
+            module = importlib.import_module(self._module)
+            self._loaded = getattr(module, self._class_name)
+        return self._loaded
+
+    def create(
+        self,
+        num_fewshot: int,
+        custom_subjects: list[str] | None,
+        custom_hf_revision: str | None,
+        user_prompt_suffix: str | None = None,
+        seed: int | None = None,
+    ) -> BaseTask:
+        return self.task_class().with_overwrite(
+            num_fewshot=num_fewshot,
+            custom_subjects=custom_subjects,
+            custom_hf_revision=custom_hf_revision,
+            user_prompt_suffix=user_prompt_suffix,
+            seed=seed,
+        )
+
+    def response_type(self) -> ResponseType:
+        """The eval's response type"""
+        return self.task_class().get_response_type()
+
+    def metrics(self) -> list[type["BaseMetric"]]:
+        """The eval's metrics"""
+        return self.task_class().get_metrics()
+
+    def subjects(self) -> list[Any]:
+        """The eval's subjects"""
+        return self.task_class().SUBJECTS
+
+    def display_name(self) -> str:
+        """The eval's human-readable display name (the task's ``NAME``)."""
+        return self.task_class().NAME
+
+    def markdown_doc(self, formatters: Sequence[BaseFormatter]) -> str:
+        try:
+            task = self.create(num_fewshot=1, custom_subjects=None, custom_hf_revision=None, seed=RANDOM_SEED)
+        except (TypeError, ValueError, AssertionError):
+            task = self.create(num_fewshot=0, custom_subjects=None, custom_hf_revision=None, seed=RANDOM_SEED)
+        return task.markdown_doc(formatters)
+
+
+class Eager(EvalFactory):
+    """Wraps an already-imported task class."""
+
+    def __init__(self, task: type[BaseTask]) -> None:
+        self._task = task
+
+    @property
+    def source_module(self) -> str:
+        return self._task.__module__
+
+    def id(self) -> str:
+        return self._task.__name__
+
+    def create(
+        self,
+        num_fewshot: int,
+        custom_subjects: list[str] | None,
+        custom_hf_revision: str | None,
+        user_prompt_suffix: str | None = None,
+        seed: int | None = None,
+    ) -> BaseTask:
+        return self._task.with_overwrite(
+            num_fewshot=num_fewshot,
+            custom_subjects=custom_subjects,
+            custom_hf_revision=custom_hf_revision,
+            user_prompt_suffix=user_prompt_suffix,
+            seed=seed,
+        )
+
+    def response_type(self) -> ResponseType:
+        """The eval's response type"""
+        return self._task.get_response_type()
+
+    def metrics(self) -> list[type["BaseMetric"]]:
+        """The eval's metrics"""
+        return self._task.get_metrics()
+
+    def subjects(self) -> list[Any]:
+        """The eval's subjects"""
+        return self._task.SUBJECTS
+
+    def display_name(self) -> str:
+        """The eval's human-readable display name (the task's ``NAME``)."""
+        return self._task.NAME
+
+    def markdown_doc(self, formatters: Sequence[BaseFormatter]) -> str:
+        try:
+            task = self.create(num_fewshot=1, custom_subjects=None, custom_hf_revision=None, seed=RANDOM_SEED)
+        except (TypeError, ValueError, AssertionError):
+            task = self.create(num_fewshot=0, custom_subjects=None, custom_hf_revision=None, seed=RANDOM_SEED)
+        return task.markdown_doc(formatters)
